--- components/index.ts
@@ -1,4 +1,0 @@
-export * as Buttons from './Buttons';
-export * as Inputs from './Inputs';
-export * as ReviewItem from './ReviewItem';
-export * as StarRating from './StarRating';
