+++ src/components/index.ts
@@ -1,4 +1,6 @@
 export * as Buttons from './Buttons';
 export * as Inputs from './Inputs';
-export * as ReviewItem from './ReviewItem';
+export * from './ReviewItem';
 export * as StarRating from './StarRating';
+export * from './TabMenu';
+export * from './PriceCalculator';
